Move nodeGetVolumeStats tests to csi_client_test.go
